Switched Decode to the new filter API.
Sorry for not answering all the emails at the moment. I am quite busy at the moment and thought that it would be more important to 
finish the code.

Craig is absolutely right with his change to auto_ptr in CreateFilter and also in regard to exceptions thrown by PoDoFo.
My next change will be to add chaining of filters, so that Encode of one Filter can write to another filter.