A large change: PdfDocument is now PdfMemDocument and both PdfMemDocument and PdfStreamedDocument implement the interface PdfDocument.
PdfMemDocument allows access to the internal object structure and keeps everything in memory.
PdfStreamedDocument was designed to not allow access to the internal object structure. It should be an easy to use class for only creating PDF 
files. It writes streams to disk (to an PdfOutputDevice) as soon as possible.

The examples where changed to either create a PdfMemDocument or PdfStreamedDocument. All applications using PdfDocument have to be changed to eihter 
construct a PdfMemDocument or a PdfStreamedDocument as a PdfDocument cannot be constructed directly.
The advantage is that all applications should be able to work with either PdfMemDocument or PdfStreamedDocument by using the PdfDocument interface 
without caring how the data is stored on disk!

PdfElements have been change so that they all can be constructed from a PdfDocument. As a reason one can use PdfElements now all with both 
PdfDocument subclasses. 

All my tests passed so it seems nothing is broken by this large change :)